basic counter works but creates 'null', going to refactor to backend increment
